# 4043 filter smt lemmas use only opaque use equations symmetrically