Related: #4034

Builds upon prior PR in #4048 

Booster SMT bindings change as follows:

* SMT lemmas are held in a lookup table separate from the sort and function declarations. The lookup is by `SymbolName`, and returns lemmas that (transitively) may relate to the given symbol
* **this PR only** This lookup table only considers _symbols with `smt-lib`_ (called _opaque_ symbols).  Symbols with `smt-hook` are left out because no lemmas should be necessary for those symbols (the solver should know all valuable lemmas about its built-in operations) 
* When checking predicates or computing a model, only those lemmas are added which relate to an _opaque_ symbol used in one of the predicates

This is hopefully reducing load on the SMT solver to enable solving problems that came up in engagement proofs and led to timeouts.