@jberthold 2024-10-03
Same picture as before for `kevm`: `sum-to-n` slowdown

| Test                                           | 4043-filter-smt-lemmas-use-only-opaque-use-equations-symmetrically time | master-4e95286ea time | (4043-filter-smt-lemmas-use-only-opaque-use-equations-symmetrically/master-4e95286ea) time 
|------------------------------------------------|-------------------------------------------------------------------------|-----------------------|--------------------------------------------------------------------------------------------
| benchmarks/ecrecoverloop02-sig1-invalid-spec.k | 65.03                                                                   | 67.57                 | 0.9624093532632826                                                                         
| benchmarks/bytes00-spec.k                      | 39.02                                                                   | 40.39                 | 0.9660807130477842                                                                         
| examples/sum-to-n-spec.k                       | 96.61                                                                   | 71.21                 | 1.356691475916304                                                                          
| TOTAL                                          | 200.66000000000003                                                      | 179.17                | 1.1199419545682874